Ownership

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
 import { DeleteResult } from 'typeorm';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,24 +15,26 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getSingleTask(id: number): Promise<Task> {
-    const foundTask = await this.taskRepository.findOne(id);
+  async getSingleTask(id: number, user: User): Promise<Task> {
+    const foundTask = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!foundTask) {
       throw new NotFoundException(`Requested Task doesn't exist`);
     }
     return foundTask;
   }
 
-  createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<DeleteResult> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<DeleteResult> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Requested Task doesn't exist`);
     }
@@ -39,11 +42,12 @@ export class TasksService {
     return result;
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    console.log({ id, status });
-
-    const task = await this.getSingleTask(id);
-    console.log(task);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getSingleTask(id, user);
 
     task.status = status;
     await task.save();
